implement scenario parser: `milestones` and `goals`
